refactor: reduce file count by consolidating related code
- Merge memory_stats.go into execution_metrics.go (both handle metrics)
- Merge streaming_processors.go into streaming.go (factory function belongs with main logic)
- Remove TODO comment from main_test.go (changed to suggestion)

This reduces the number of files from 31 to 29 while maintaining logical organization.
