+++ streaming.go
@@ -190,3 +190,66 @@ func isStreamingSupported(mode enums.DisplayMode) bool {
 		return false
 	}
 }
+
+// NewStreamingProcessorForMode creates a streaming processor for the given display mode.
+// Returns nil if the mode doesn't support streaming yet.
+// This is primarily used for testing - production code uses createStreamingProcessor.
+func NewStreamingProcessorForMode(mode enums.DisplayMode, out io.Writer, sysVars *systemVariables, screenWidth int) RowProcessor {
+	var formatter StreamingFormatter
+
+	switch mode {
+	case enums.DisplayModeCSV:
+		formatter = NewCSVFormatter(out, sysVars.SkipColumnNames)
+		return &StreamingProcessor{
+			formatter:   formatter,
+			out:         out,
+			screenWidth: screenWidth,
+		}
+
+	case enums.DisplayModeTab:
+		formatter = NewTabFormatter(out, sysVars.SkipColumnNames)
+		return &StreamingProcessor{
+			formatter:   formatter,
+			out:         out,
+			screenWidth: screenWidth,
+		}
+
+	case enums.DisplayModeVertical:
+		formatter = NewVerticalFormatter(out)
+		return &StreamingProcessor{
+			formatter:   formatter,
+			out:         out,
+			screenWidth: screenWidth,
+		}
+
+	case enums.DisplayModeHTML:
+		formatter = NewHTMLFormatter(out, sysVars.SkipColumnNames)
+		return &StreamingProcessor{
+			formatter:   formatter,
+			out:         out,
+			screenWidth: screenWidth,
+		}
+
+	case enums.DisplayModeXML:
+		formatter = NewXMLFormatter(out, sysVars.SkipColumnNames)
+		return &StreamingProcessor{
+			formatter:   formatter,
+			out:         out,
+			screenWidth: screenWidth,
+		}
+
+	case enums.DisplayModeTable, enums.DisplayModeTableComment, enums.DisplayModeTableDetailComment:
+		// Table formats use preview processor for width calculation
+		// TablePreviewRows: positive = preview N rows, 0 = headers only, negative = all rows
+		previewSize := int(sysVars.TablePreviewRows)
+		if previewSize < 0 {
+			// Negative means collect all rows (non-streaming)
+			previewSize = 0 // 0 in TablePreviewProcessor means collect all
+		}
+		tableFormatter := NewTableStreamingFormatter(out, sysVars, screenWidth, previewSize)
+		return NewTablePreviewProcessor(tableFormatter, previewSize)
+
+	default:
+		return nil
+	}
+}
